Cut 8 bytes out of LSPFileUpdates::FastPathjfilesToTypecheckResult (#10287)

+++ main/lsp/LSPFileUpdates.h
@@ -77,7 +77,7 @@ public:
 
     struct FastPathFilesToTypecheckResult {
         // The number of files that would be checked in the fast path.
-        size_t totalChanged = 0;
+        uint32_t totalChanged = 0;
 
         // True when we should use the incremental namer, which happens if a symbol name changed and we brought in
         // additional related files to check.
